Organize renderer configuration and fix documentation CI
