+++ examples/simulation/articulated/simulate_articulated_soft_robot.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import jax
 
-from soromox.rendering.renderer_config import GeometryConfig, RendererConfig
+from soromox.rendering.config import GeometryConfig, RendererConfig
 
 jax.config.update("jax_enable_x64", True)  # double precision
