+++ examples/simulation/pcs/simulate_batched_tendon_actuated_pcs.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 import jax
 
-from soromox.rendering.renderer_config import GeometryConfig, RendererConfig
+from soromox.rendering.config import GeometryConfig, RendererConfig
 
 jax.config.update("jax_enable_x64", True)
 import jax.numpy as jnp
