+++ examples/simulation/pcs/simulate_tendon_actuated_pcs.py
@@ -21,7 +21,7 @@ from soromox.rendering import (
     RendererColorConfig,
     ViserRenderer,
 )
-from soromox.rendering.renderer_config import GeometryConfig, RendererConfig
+from soromox.rendering.config import GeometryConfig, RendererConfig
 from soromox.systems import (
     PCS,
     LinkSpec,
