+++ paper_results/secVc_model_based_control/configuration_space_comparison/code/render_configuration_space_comparison.py
@@ -18,7 +18,7 @@ from configuration_space_comparison_simulation import (
 )
 
 from soromox.rendering import BackboneColorConfig, RendererColorConfig, ViserRenderer
-from soromox.rendering.renderer_config import GeometryConfig, RendererConfig
+from soromox.rendering.config import GeometryConfig, RendererConfig
 
 ACTUAL_RENDER_COLORS = np.array(
     [
