+++ paper_results/secVe_safety_constrained_control/code/render_pcs_with_cf_cbf_clf_video.py
@@ -24,7 +24,7 @@ from soromox.rendering import (
     Open3DRenderer,
     RendererColorConfig,
 )
-from soromox.rendering.renderer_config import (
+from soromox.rendering.config import (
     GeometryConfig,
     RendererConfig,
     RenderOutputConfig,
