+++ src/soromox/rendering/config/camera.py
