+++ src/soromox/rendering/config/scene.py
@@ -1,63 +1,18 @@
-"""Backend-independent renderer settings with photometric light units."""
+"""Scene, lighting, material and ground settings with photometric units."""
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field, fields, is_dataclass, replace
+from dataclasses import dataclass, field, replace
 from typing import Literal
 
 import numpy as np
 
-from soromox.rendering.camera_config import CameraConfig
-from soromox.rendering.color_config import (
-    ActuatorStyleConfig,
-    BackboneColorConfig,
-    RendererColorConfig,
+from soromox.rendering.config._validation import (
+    RGB,
+    Vector3,
+    _nonnegative,
+    _validate_vector,
 )
-from soromox.rendering.video_encoding import VideoEncodingConfig
-
-RGB = tuple[float, float, float]
-Vector3 = tuple[float, float, float]
-
-
-def _validate_vector(value, name, *, color=False, nonzero=False):
-    """Validate a finite vector, optionally restricting its range or length.
-
-    Args:
-        value: Three numeric components.
-        name: Field name used in errors.
-        color: Require sRGB components in [0, 1].
-        nonzero: Require a nonzero vector.
-
-    Returns:
-        None.
-
-    Raises:
-        ValueError: The vector violates a requested constraint.
-    """
-    a = np.asarray(value, dtype=float)
-    if a.shape != (3,) or not np.isfinite(a).all():
-        raise ValueError(f"{name} must be a finite three-vector")
-    if color and ((a < 0).any() or (a > 1).any()):
-        raise ValueError(f"{name} must lie in [0, 1]")
-    if nonzero and np.linalg.norm(a) == 0:
-        raise ValueError(f"{name} must be nonzero")
-
-
-def _nonnegative(value, name):
-    """Validate a finite nonnegative scalar.
-
-    Args:
-        value: Numeric scalar.
-        name: Field name used in errors.
-
-    Returns:
-        None.
-
-    Raises:
-        ValueError: The value is negative or nonfinite.
-    """
-    if not np.isfinite(value) or value < 0:
-        raise ValueError(f"{name} must be finite and nonnegative")
 
 
 @dataclass
@@ -484,206 +439,3 @@ class SceneConfig:
             ),
             **overrides,
         )
-
-
-@dataclass
-class GeometryConfig:
-    """Shared geometry sampling and display dimensions.
-
-    Attributes:
-        num_points: Backbone samples, at least two.
-        cross_section_resolution: Contour samples, at least three.
-        backbone_style: Swept surface or discrete cross-section markers.
-        base_plate_radius_scale: Base plate radius relative to cross-section size.
-        base_plate_thickness: Plate thickness in metres.
-        line_width: Backbone width in pixels for line-based renderers; None selects
-            a cross-section-derived width where supported.
-        actuator_line_width: Actuator line width in pixels.
-        grid_spacing: Multi-robot layout spacing in metres.
-    """
-
-    num_points: int = 80
-    cross_section_resolution: int = 48
-    backbone_style: Literal["swept", "discrete"] = "swept"
-    base_plate_radius_scale: float = 2.0
-    base_plate_thickness: float = 0.06
-    line_width: float | None = 4.0
-    actuator_line_width: float = 2.0
-    grid_spacing: tuple[float, float] = (0.5, 0.5)
-
-    def __post_init__(self):
-        """Validate sample counts, dimensions and backbone style.
-
-        Returns:
-            None.
-
-        Raises:
-            ValueError: A count, dimension or style is invalid.
-        """
-        for name, minimum in (("num_points", 2), ("cross_section_resolution", 3)):
-            value = getattr(self, name)
-            if (
-                isinstance(value, bool)
-                or not isinstance(value, (int, np.integer))
-                or value < minimum
-            ):
-                raise ValueError(f"{name} must be an integer >= {minimum}")
-        if self.backbone_style not in ("swept", "discrete"):
-            raise ValueError("backbone_style must be swept or discrete")
-        for name in (
-            "base_plate_radius_scale",
-            "base_plate_thickness",
-            "actuator_line_width",
-        ):
-            _nonnegative(getattr(self, name), name)
-        if self.line_width is not None:
-            _nonnegative(self.line_width, "line_width")
-        spacing = np.asarray(self.grid_spacing)
-        if (
-            spacing.shape != (2,)
-            or not np.isfinite(spacing).all()
-            or (spacing <= 0).any()
-        ):
-            raise ValueError("grid_spacing must contain two positive finite distances")
-
-
-@dataclass
-class RenderOutputConfig:
-    """Image dimensions and video encoding defaults.
-
-    Attributes:
-        width: Positive image width in pixels.
-        height: Positive image height in pixels.
-        video: Default video encoding configuration.
-    """
-
-    width: int = 800
-    height: int = 600
-    video: VideoEncodingConfig = field(default_factory=VideoEncodingConfig)
-
-    def __post_init__(self):
-        """Validate integer image dimensions.
-
-        Returns:
-            None.
-
-        Raises:
-            ValueError: A dimension is not a positive integer.
-        """
-        for name in ("width", "height"):
-            value = getattr(self, name)
-            if (
-                isinstance(value, bool)
-                or not isinstance(value, (int, np.integer))
-                or value <= 0
-            ):
-                raise ValueError(f"{name} must be a positive integer")
-
-
-@dataclass
-class RendererConfig:
-    """Composed configuration accepted by every renderer.
-
-    Attributes:
-        scene: Scene appearance, lighting and ground plane.
-        camera: Default camera and exposure.
-        colors: Robot colors; helper colors are supplied to rendering operations.
-        geometry: Sampling, mesh style and display dimensions.
-        output: Image dimensions and video encoding.
-    """
-
-    scene: SceneConfig = field(default_factory=SceneConfig.technical)
-    camera: CameraConfig = field(default_factory=CameraConfig)
-    colors: RendererColorConfig = field(default_factory=RendererColorConfig)
-    geometry: GeometryConfig = field(default_factory=GeometryConfig)
-    output: RenderOutputConfig = field(default_factory=RenderOutputConfig)
-
-    @classmethod
-    def clay(
-        cls, color: RGB = (0.72, 0.65, 0.56), *, scene_extent: float = 1.2, **overrides
-    ) -> RendererConfig:
-        """Create a studio with uniform matte clay robot surfaces.
-
-        Args:
-            color: Warm-grey sRGB clay color; helper object colors are unaffected.
-            scene_extent: Reference scene extent in metres for preset lighting.
-            **overrides: Complete configuration sections replacing preset defaults.
-
-        Returns:
-            Editable renderer configuration with monochrome robot colors.
-
-        Raises:
-            ValueError: Color or scene extent is invalid.
-        """
-        _validate_vector(color, "color", color=True)
-        scene = SceneConfig.studio(scene_extent=scene_extent)
-        scene.material.roughness = 0.9
-        scale = scene_extent / 1.2
-        scene.lights += (
-            PointLightConfig.from_lumens(
-                100000 * scale**2,
-                position=tuple(np.array([0.7, 0.4, 0.6]) * scale),
-                range_m=6.0 * scale,
-                color=(0.85, 0.92, 1.0),
-            ),
-        )
-        colors = RendererColorConfig(
-            backbone=BackboneColorConfig(
-                segment_palette=[color], robot_palette=[color]
-            ),
-            base_plate_color=color,
-            actuators=ActuatorStyleConfig(default_color=color),
-            robot_override=color,
-        )
-        return replace(cls(scene=scene, colors=colors), **overrides)
-
-
-def validate_config(config: RendererConfig) -> None:
-    """Validate an editable configuration immediately before renderer construction.
-
-    Args:
-        config: Complete renderer configuration, including edits made after a
-            preset factory returned it.
-
-    Returns:
-        None. Validates nested dataclass fields without changing them.
-
-    Raises:
-        TypeError: A top-level section has an incorrect type.
-        ValueError: A field violates its configuration constraints.
-    """
-    for name, kind in (
-        ("scene", SceneConfig),
-        ("camera", CameraConfig),
-        ("colors", RendererColorConfig),
-        ("geometry", GeometryConfig),
-        ("output", RenderOutputConfig),
-    ):
-        if not isinstance(getattr(config, name), kind):
-            raise TypeError(f"{name} must be {kind.__name__}")
-
-    def visit(value):
-        """Validate nested values before their containing dataclass.
-
-        Args:
-            value: A configuration value or collection.
-
-        Returns:
-            None.
-
-        Raises:
-            ValueError: A nested configuration fails validation.
-        """
-        if is_dataclass(value):
-            for item in fields(value):
-                visit(getattr(value, item.name))
-            validator = getattr(value, "__post_init__", None)
-            if validator is not None:
-                validator()
-        elif isinstance(value, (list, tuple)):
-            for item in value:
-                visit(item)
-
-    visit(config)
-    if not np.isfinite(config.camera.exposure_ev100):
-        raise ValueError("exposure_ev100 must be finite")
