+++ tests/rendering/test_viser_renderer.py
@@ -9,14 +9,14 @@ from numpy.testing import assert_allclose, assert_array_equal
 from PIL import Image
 
 from soromox.rendering.actuators import ActuatorVisualLayer
-from soromox.rendering.camera_config import CameraConfig
-from soromox.rendering.renderer_config import (
+from soromox.rendering.config import (
     GeometryConfig,
     GroundPlaneConfig,
     RendererConfig,
     RenderOutputConfig,
     SceneConfig,
 )
+from soromox.rendering.config.camera import CameraConfig
 from soromox.systems.components import CrossSectionGeometry
 from soromox.utils.geometry import poses
