+++ tests/systems/test_isupport.py
@@ -3,7 +3,7 @@
 # ruff: noqa: E402
 import jax
 
-from soromox.rendering.renderer_config import GeometryConfig, RendererConfig
+from soromox.rendering.config import GeometryConfig, RendererConfig
 
 jax.config.update("jax_enable_x64", True)
